FinStep Class,MIOpenJobMixin(rename) update
