+++ tuna/miopen/db/miopen_tables.py
@@ -43,7 +43,6 @@ from tuna.miopen.db.session import Session
 from tuna.miopen.utils.metadata import DIR_MAP
 from tuna.miopen.db.benchmark import Model, Framework
 from tuna.db.tuna_tables import JobMixin
-from tuna.db.tuna_tables import FinStep
 
 COMMON_UNIQ_FDS = ["config", "solver", "session"]
 
@@ -309,28 +308,6 @@ class FusionConfigTags(BASE, ConfigTagMixin):
   config = Column(Integer, ForeignKey("fusion_config.id"), nullable=False)
 
 
-class MIOpenJob(BASE, JobMixin):
-  """Represents MIOpen Mixin class for job tables"""
-
-  @declared_attr
-  def session(self):
-    """session key"""
-    return Column(Integer, ForeignKey("session.id"), nullable=False)
-
-  compile_start = Column(DateTime,
-                         nullable=False,
-                         server_default=sqla_func.now())
-  compile_end = Column(DateTime, nullable=False, server_default=sqla_func.now())
-  eval_start = Column(DateTime, nullable=False, server_default=sqla_func.now())
-  eval_end = Column(DateTime, nullable=False, server_default=sqla_func.now())
-
-  solver = Column(String(length=128), nullable=True, server_default="")
-  eval_mid = Column(Integer, server_default="-1")
-  fin_step = Column(mysql.MSSet(*(list(k for k in FinStep.__members__))),
-                    nullable=False,
-                    server_default="not_fin")
-
-
 class JobEnum(enum.Enum):
   """Represents job_enum column in config table"""
   # pylint: disable=invalid-name ; names represent entries in job_enum column
@@ -361,7 +338,43 @@ class JobEnum(enum.Enum):
   evaluated_pend = 24
 
 
-class ConvolutionJob(BASE, MIOpenJob):
+class FinStep(enum.Enum):
+  """ Allowed Fin Steps """
+  # pylint: disable=invalid-name ; tuna/go_fish.py names valid fin steps as FinStep.__members__
+  find_compile = 1
+  find_eval = 2
+  get_solvers = 3
+  get_applicability = 4
+  not_fin = 5
+  miopen_find_compile = 6
+  miopen_find_eval = 7
+  miopen_perf_compile = 8
+  miopen_perf_eval = 9
+
+
+class MIOpenJobMixin(BASE, JobMixin):
+  """Represents MIOpen Mixin class for job tables"""
+
+  @declared_attr
+  def session(self):
+    """session key"""
+    return Column(Integer, ForeignKey("session.id"), nullable=False)
+
+  compile_start = Column(DateTime,
+                         nullable=False,
+                         server_default=sqla_func.now())
+  compile_end = Column(DateTime, nullable=False, server_default=sqla_func.now())
+  eval_start = Column(DateTime, nullable=False, server_default=sqla_func.now())
+  eval_end = Column(DateTime, nullable=False, server_default=sqla_func.now())
+
+  solver = Column(String(length=128), nullable=True, server_default="")
+  eval_mid = Column(Integer, server_default="-1")
+  fin_step = Column(mysql.MSSet(*(list(k for k in FinStep.__members__))),
+                    nullable=False,
+                    server_default="not_fin")
+
+
+class ConvolutionJob(MIOpenJobMixin):
   """Represents convolutions job table"""
   __tablename__ = "conv_job"
   __table_args__ = (UniqueConstraint(*COMMON_UNIQ_FDS, name="uq_idx"),)
@@ -378,7 +391,7 @@ class ConvolutionJob(BASE, MIOpenJob):
                           'session')
 
 
-class BNJob(BASE, MIOpenJob):
+class BNJob(MIOpenJobMixin):
   """Represents batch norm job table"""
   __tablename__ = "bn_job"
   __table_args__ = (UniqueConstraint(*COMMON_UNIQ_FDS, name="uq_idx"),)
@@ -389,7 +402,7 @@ class BNJob(BASE, MIOpenJob):
                   index=True)
 
 
-class FusionJob(BASE, MIOpenJob):
+class FusionJob(MIOpenJobMixin):
   """Represents fusions job table"""
   __tablename__ = "fusion_job"
   __table_args__ = (UniqueConstraint(*COMMON_UNIQ_FDS, name="uq_idx"),)
